# improve sampling time