Summary:
Before the change
```
I 00:00:01.193511 executorch:runner.cpp:145] sampling took 47.523000 ms
I 00:00:01.193581 executorch:runner.cpp:230] t2_1-t1 took 48.221000 ms
bageI 00:00:01.193678 executorch:runner.cpp:237] t3-t2 took 47.716000 ms
I 00:00:01.193719 executorch:runner.cpp:242] t4-t3 took 0.041000 ms
I 00:00:01.193751 executorch:runner.cpp:248] t5-t4 took 0.032000 ms
I 00:00:01.193765 executorch:runner.cpp:449] prefill took 116.108000 ms
I 00:00:01.215019 executorch:runner.cpp:145] sampling took 21.223000 ms
```

After the change
```
I 00:00:01.295467 executorch:runner.cpp:145] sampling took 36.823000 ms
I 00:00:01.295528 executorch:runner.cpp:230] t2_1-t1 took 37.779000 ms
 I 00:00:01.295577 executorch:runner.cpp:237] t3-t2 took 36.981000 ms
I 00:00:01.295600 executorch:runner.cpp:242] t4-t3 took 0.023000 ms
I 00:00:01.295621 executorch:runner.cpp:248] t5-t4 took 0.020000 ms
I 00:00:01.295647 executorch:runner.cpp:449] prefill took 98.039000 ms
I 00:00:01.311898 executorch:runner.cpp:145] sampling took 16.217000 ms
```

before the change, end to end latency is 149 ms:
```
I 00:00:03.632890 executorch:runner.cpp:74] Reading metadata from model
I 00:00:03.633021 executorch:runner.cpp:128] get_n_bos: 1
I 00:00:03.633064 executorch:runner.cpp:128] get_n_eos: 1
I 00:00:03.633080 executorch:runner.cpp:128] use_kv_cache: 0
I 00:00:03.633094 executorch:runner.cpp:128] use_sdpa_with_kv_cache: 0
I 00:00:03.633106 executorch:runner.cpp:128] append_eos_to_prompt: 0
I 00:00:03.633118 executorch:runner.cpp:128] enable_dynamic_shape: 0
I 00:00:03.991977 executorch:runner.cpp:128] get_vocab_size: 128256
I 00:00:03.992212 executorch:runner.cpp:128] get_bos_id: 1
I 00:00:03.992269 executorch:runner.cpp:128] get_eos_id: 2
I 00:00:04.019936 executorch:runner.cpp:369] encoder took 27.617000 ms
[WARNING] [Qnn ExecuTorch]:  <W> sg_stubPtr is not null, skip loadRemoteSymbols

I 00:00:04.064693 executorch:runner.cpp:186] inference took 41.274000 ms
How does artificial intelligence redefine the role of human creativity in the next decade? compassionI 00:00:04.145481 executorch:stats.h:70]      Prompt Tokens: 16    Generated Tokens: 0
I 00:00:04.145578 executorch:stats.h:76]        Model Load Time:                3.985000 (seconds)
I 00:00:04.145619 executorch:stats.h:86]        Total inference time:           0.149000 (seconds)               Rate:  0.000000 (tokens/second)
I 00:00:04.145654 executorch:stats.h:94]                Prompt evaluation:      0.128000 (seconds)               Rate:  125.000000 (tokens/second)
I 00:00:04.145688 executorch:stats.h:105]               Generated 0 tokens:     0.021000 (seconds)               Rate:  0.000000 (tokens/second)
I 00:00:04.145721 executorch:stats.h:113]       Time to first generated token:  0.128000 (seconds)
I 00:00:04.145754 executorch:stats.h:120]       Sampling time over 16 tokens:   0.000000 (seconds)
[INFO] [Qnn ExecuTorch]: Destroy Qnn backend parameters
[INFO] [Qnn ExecuTorch]: Destroy Qnn context
[WARNING] [Qnn ExecuTorch]:  <W> sg_stubPtr is not null, skip loadRemoteSymbols

```

after the change, end to end latency 116 ms
```
[INFO] [Qnn ExecuTorch]: Running level=3 optimization.
I 00:00:00.687492 executorch:runner.cpp:74] Reading metadata from model
I 00:00:00.687603 executorch:runner.cpp:128] get_n_bos: 1
I 00:00:00.687654 executorch:runner.cpp:128] get_n_eos: 1
I 00:00:00.687697 executorch:runner.cpp:128] use_kv_cache: 0
I 00:00:00.687738 executorch:runner.cpp:128] use_sdpa_with_kv_cache: 0
I 00:00:00.687777 executorch:runner.cpp:128] append_eos_to_prompt: 0
I 00:00:00.687817 executorch:runner.cpp:128] enable_dynamic_shape: 0
I 00:00:01.037239 executorch:runner.cpp:128] get_vocab_size: 128256
I 00:00:01.037451 executorch:runner.cpp:128] get_bos_id: 1
I 00:00:01.037502 executorch:runner.cpp:128] get_eos_id: 2
I 00:00:01.061254 executorch:runner.cpp:369] encoder took 23.705000 ms
[WARNING] [Qnn ExecuTorch]:  <W> sg_stubPtr is not null, skip loadRemoteSymbols

I 00:00:01.097999 executorch:runner.cpp:186] inference took 36.617000 ms
How does artificial intelligence redefine the role of human creativity in the next decade?;'>I 00:00:01.153426 executorch:stats.h:70]      Prompt Tokens: 16    Generated Tokens: 0
I 00:00:01.153532 executorch:stats.h:76]        Model Load Time:                1.032000 (seconds)
I 00:00:01.153573 executorch:stats.h:86]        Total inference time:           0.116000 (seconds)               Rate:  0.000000 (tokens/second)
I 00:00:01.153612 executorch:stats.h:94]                Prompt evaluation:      0.099000 (seconds)               Rate:  161.616162 (tokens/second)
I 00:00:01.153668 executorch:stats.h:105]               Generated 0 tokens:     0.017000 (seconds)               Rate:  0.000000 (tokens/second)
I 00:00:01.153706 executorch:stats.h:113]       Time to first generated token:  0.099000 (seconds)
I 00:00:01.153739 executorch:stats.h:120]       Sampling time over 16 tokens:   0.000000 (seconds)
[INFO] [Qnn ExecuTorch]: Destroy Qnn backend parameters
```

Reviewed By: swolchok

Differential Revision: D60742125